Android 14.0.0 release 29

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCZeZW4QAKCRDorT+BmrEO
# eE4/AJ9E/RZ0j1nWQ520q2J/M7xvr7x9bwCeIbsmQ5h/VhK1tESzl8+Xygmrv0M=
# =N+BS
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue Mar  5 07:18:57 2024 CST
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Can't check signature: No public key

# Conflicts:
#	build.gradle
#	quickstep/res/layout/split_instructions_view.xml
#	quickstep/res/layout/task_desktop.xml
#	quickstep/res/values-de/strings.xml
#	quickstep/src/com/android/launcher3/QuickstepTransitionManager.java
#	quickstep/src/com/android/launcher3/model/QuickstepModelDelegate.java
#	quickstep/src/com/android/launcher3/taskbar/NavbarButtonsViewController.java
#	quickstep/src/com/android/launcher3/taskbar/TaskbarActivityContext.java
#	quickstep/src/com/android/launcher3/taskbar/TaskbarLauncherStateController.java
#	quickstep/src/com/android/launcher3/taskbar/TaskbarManager.java
#	quickstep/src/com/android/launcher3/taskbar/TaskbarPopupController.java
#	quickstep/src/com/android/launcher3/taskbar/TaskbarStashController.java
#	quickstep/src/com/android/launcher3/taskbar/TaskbarUIController.java
#	quickstep/src/com/android/launcher3/taskbar/TaskbarView.java
#	quickstep/src/com/android/launcher3/taskbar/allapps/TaskbarAllAppsController.java
#	quickstep/src/com/android/launcher3/taskbar/bubbles/BubbleBarController.java
#	quickstep/src/com/android/launcher3/taskbar/navbutton/AbstractNavButtonLayoutter.kt
#	quickstep/src/com/android/launcher3/uioverrides/ApiWrapper.java
#	quickstep/src/com/android/launcher3/uioverrides/QuickstepLauncher.java
#	quickstep/src/com/android/launcher3/uioverrides/touchcontrollers/NoButtonQuickSwitchTouchController.java
#	quickstep/src/com/android/launcher3/uioverrides/touchcontrollers/StatusBarTouchController.java
#	quickstep/src/com/android/quickstep/AbsSwipeUpHandler.java
#	quickstep/src/com/android/quickstep/LauncherBackAnimationController.java
#	quickstep/src/com/android/quickstep/RecentsActivity.java
#	quickstep/src/com/android/quickstep/RecentsAnimationDeviceState.java
#	quickstep/src/com/android/quickstep/RecentsModel.java
#	quickstep/src/com/android/quickstep/SystemUiProxy.java
#	quickstep/src/com/android/quickstep/TaskAnimationManager.java
#	quickstep/src/com/android/quickstep/TaskIconCache.java
#	quickstep/src/com/android/quickstep/TaskOverlayFactory.java
#	quickstep/src/com/android/quickstep/TaskShortcutFactory.java
#	quickstep/src/com/android/quickstep/TaskViewUtils.java
#	quickstep/src/com/android/quickstep/TouchInteractionService.java
#	quickstep/src/com/android/quickstep/inputconsumers/TaskbarUnstashInputConsumer.java
#	quickstep/src/com/android/quickstep/util/AppPairsController.java
#	quickstep/src/com/android/quickstep/util/AsyncClockEventDelegate.java
#	quickstep/src/com/android/quickstep/util/BorderAnimator.java
#	quickstep/src/com/android/quickstep/util/TaskViewSimulator.java
#	quickstep/src/com/android/quickstep/util/TransformParams.java
#	quickstep/src/com/android/quickstep/views/DigitalWellBeingToast.java
#	quickstep/src/com/android/quickstep/views/GroupedTaskView.java
#	quickstep/src/com/android/quickstep/views/RecentsView.java
#	quickstep/src/com/android/quickstep/views/TaskMenuView.java
#	quickstep/src/com/android/quickstep/views/TaskView.java
#	res/values-ar/strings.xml
#	res/values-iw/strings.xml
#	res/values-ja/strings.xml
#	res/values-pt/strings.xml
#	res/values-ro/strings.xml
#	res/values-ru/strings.xml
#	res/values-zh-rCN/strings.xml
#	res/values/dimens.xml
#	src/com/android/launcher3/AbstractFloatingView.java
#	src/com/android/launcher3/AppWidgetResizeFrame.java
#	src/com/android/launcher3/BaseDraggingActivity.java
#	src/com/android/launcher3/BubbleTextView.java
#	src/com/android/launcher3/CellLayout.java
#	src/com/android/launcher3/DeviceProfile.java
#	src/com/android/launcher3/ExtendedEditText.java
#	src/com/android/launcher3/FastScrollRecyclerView.java
#	src/com/android/launcher3/Hotseat.java
#	src/com/android/launcher3/InvariantDeviceProfile.java
#	src/com/android/launcher3/Launcher.java
#	src/com/android/launcher3/LauncherAppState.java
#	src/com/android/launcher3/LauncherProvider.java
#	src/com/android/launcher3/ShortcutAndWidgetContainer.java
#	src/com/android/launcher3/Utilities.java
#	src/com/android/launcher3/Workspace.java
#	src/com/android/launcher3/allapps/ActivityAllAppsContainerView.java
#	src/com/android/launcher3/allapps/AllAppsStore.java
#	src/com/android/launcher3/allapps/AllAppsTransitionController.java
#	src/com/android/launcher3/allapps/BaseAllAppsAdapter.java
#	src/com/android/launcher3/allapps/WorkProfileManager.java
#	src/com/android/launcher3/config/FeatureFlags.java
#	src/com/android/launcher3/dragndrop/DragView.java
#	src/com/android/launcher3/folder/Folder.java
#	src/com/android/launcher3/folder/LauncherDelegate.java
#	src/com/android/launcher3/folder/PreviewItemManager.java
#	src/com/android/launcher3/graphics/LauncherPreviewRenderer.java
#	src/com/android/launcher3/graphics/SysUiScrim.java
#	src/com/android/launcher3/icons/LauncherIcons.java
#	src/com/android/launcher3/logging/StartupLatencyLogger.kt
#	src/com/android/launcher3/logging/StatsLogManager.java
#	src/com/android/launcher3/model/BaseLauncherBinder.java
#	src/com/android/launcher3/model/BgDataModel.java
#	src/com/android/launcher3/model/DatabaseHelper.java
#	src/com/android/launcher3/model/GridSizeMigrationUtil.java
#	src/com/android/launcher3/model/LoaderCursor.java
#	src/com/android/launcher3/model/LoaderTask.java
#	src/com/android/launcher3/model/ModelDbController.java
#	src/com/android/launcher3/model/ModelWriter.java
#	src/com/android/launcher3/model/PackageUpdatedTask.java
#	src/com/android/launcher3/model/WorkspaceItemSpaceFinder.java
#	src/com/android/launcher3/pageindicators/PageIndicatorDots.java
#	src/com/android/launcher3/pm/UserCache.java
#	src/com/android/launcher3/popup/ArrowPopup.java
#	src/com/android/launcher3/popup/PopupContainerWithArrow.java
#	src/com/android/launcher3/provider/RestoreDbTask.java
#	src/com/android/launcher3/qsb/QsbContainerView.java
#	src/com/android/launcher3/secondarydisplay/SecondaryDragLayer.java
#	src/com/android/launcher3/settings/SettingsActivity.java
#	src/com/android/launcher3/shortcuts/DeepShortcutTextView.java
#	src/com/android/launcher3/touch/AllAppsSwipeController.java
#	src/com/android/launcher3/util/DisplayController.java
#	src/com/android/launcher3/util/OnboardingPrefs.java
#	src/com/android/launcher3/util/VibratorWrapper.java
#	src/com/android/launcher3/views/ActivityContext.java
#	src/com/android/launcher3/views/FloatingIconView.java
#	src/com/android/launcher3/views/OptionsPopupView.java
#	src/com/android/launcher3/widget/BaseWidgetSheet.java
#	src/com/android/launcher3/widget/picker/WidgetsFullSheet.java
#	tests/shared/com/android/launcher3/testing/shared/TestProtocol.java
#	tests/src/com/android/launcher3/model/GridSizeMigrationUtilTest.kt
#	tests/tapl/com/android/launcher3/tapl/LauncherInstrumentation.java

+++ src/com/android/launcher3/model/DatabaseHelper.java
@@ -16,6 +16,7 @@
 package com.android.launcher3.model;
 
 import static com.android.launcher3.LauncherSettings.Favorites.addTableToDb;
+import static com.android.launcher3.config.FeatureFlags.shouldShowFirstPageWidget;
 import static com.android.launcher3.provider.LauncherDbUtils.dropTable;
 
 import android.content.ContentValues;
@@ -66,10 +67,8 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
         LayoutParserCallback {
 
     /**
-     * Represents the schema of the database. Changes in scheme need not be
-     * backwards compatible.
-     * When increasing the scheme version, ensure that downgrade_schema.json is
-     * updated
+     * Represents the schema of the database. Changes in scheme need not be backwards compatible.
+     * When increasing the scheme version, ensure that downgrade_schema.json is updated
      */
     public static final int SCHEMA_VERSION = 32;
     private static final String TAG = "DatabaseHelper";
@@ -96,8 +95,7 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
     }
 
     protected void initIds() {
-        // In the case where neither onCreate nor onUpgrade gets called, we read the
-        // maxId from
+        // In the case where neither onCreate nor onUpgrade gets called, we read the maxId from
         // the DB here
         if (mMaxItemId == -1) {
             mMaxItemId = initializeMaxItemId(getWritableDatabase());
@@ -106,8 +104,7 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        if (LOGD)
-            Log.d(TAG, "creating new launcher database");
+        if (LOGD) Log.d(TAG, "creating new launcher database");
 
         mMaxItemId = 1;
 
@@ -141,11 +138,9 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
     }
 
     /**
-     * One-time data updated before support of onDowngrade was added. This update is
-     * backwards
+     * One-time data updated before support of onDowngrade was added. This update is backwards
      * compatible and can safely be run multiple times.
-     * Note: No new logic should be added here after release, as the new logic might
-     * not get
+     * Note: No new logic should be added here after release, as the new logic might not get
      * executed on an existing device.
      * TODO: Move this to db upgrade path, once the downgrade path is released.
      */
@@ -263,7 +258,8 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
                         Favorites.SCREEN, IntArray.wrap(-777, -778)), null);
             }
             case 30: {
-                if (FeatureFlags.topQsbOnFirstScreenEnabled(mContext)) {
+                if (FeatureFlags.topQsbOnFirstScreenEnabled(mContext)
+                        && !shouldShowFirstPageWidget()) {
                     // Clean up first row in screen 0 as it might contain junk data.
                     Log.d(TAG, "Cleaning up first row");
                     db.delete(Favorites.TABLE_NAME,
@@ -271,10 +267,8 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
                                     "%1$s = %2$d AND %3$s = %4$d AND %5$s = %6$d",
                                     Favorites.SCREEN, 0,
                                     Favorites.CONTAINER, Favorites.CONTAINER_DESKTOP,
-                                    Favorites.CELLY, 0),
-                            null);
+                                    Favorites.CELLY, 0), null);
                 }
-                return;
             }
             case 31: {
                 LauncherDbUtils.migrateLegacyShortcuts(mContext, db);
@@ -316,8 +310,7 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
     }
 
     /**
-     * Removes widgets which are registered to the Launcher's host, but are not
-     * present
+     * Removes widgets which are registered to the Launcher's host, but are not present
      * in our model.
      */
     public void removeGhostWidgets(SQLiteDatabase db) {
@@ -365,21 +358,21 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
     }
 
     /**
-     * Replaces all shortcuts of type {@link Favorites#ITEM_TYPE_SHORTCUT} which
-     * have a valid
+     * Replaces all shortcuts of type {@link Favorites#ITEM_TYPE_SHORTCUT} which have a valid
      * launcher activity target with {@link Favorites#ITEM_TYPE_APPLICATION}.
      */
     @Thunk
     void convertShortcutsToLauncherActivities(SQLiteDatabase db) {
         try (SQLiteTransaction t = new SQLiteTransaction(db);
-                // Only consider the primary user as other users can't have a shortcut.
-                Cursor c = db.query(Favorites.TABLE_NAME,
-                        new String[] { Favorites._ID, Favorites.INTENT },
-                        "itemType=" + Favorites.ITEM_TYPE_SHORTCUT
-                                + " AND profileId=" + getDefaultUserSerial(),
-                        null, null, null, null);
-                SQLiteStatement updateStmt = db.compileStatement("UPDATE favorites SET itemType="
-                        + Favorites.ITEM_TYPE_APPLICATION + " WHERE _id=?")) {
+             // Only consider the primary user as other users can't have a shortcut.
+             Cursor c = db.query(Favorites.TABLE_NAME,
+                     new String[]{Favorites._ID, Favorites.INTENT},
+                     "itemType=" + Favorites.ITEM_TYPE_SHORTCUT
+                             + " AND profileId=" + getDefaultUserSerial(),
+                     null, null, null, null);
+             SQLiteStatement updateStmt = db.compileStatement("UPDATE favorites SET itemType="
+                     + Favorites.ITEM_TYPE_APPLICATION + " WHERE _id=?")
+        ) {
             final int idIndex = c.getColumnIndexOrThrow(Favorites._ID);
             final int intentIndex = c.getColumnIndexOrThrow(Favorites.INTENT);
 
@@ -417,14 +410,14 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
 
             // Get a map for folder ID to folder width
             Cursor c = db.rawQuery("SELECT container, MAX(cellX) FROM favorites"
-                    + " WHERE container IN (SELECT _id FROM favorites WHERE itemType = ?)"
-                    + " GROUP BY container;",
-                    new String[] { Integer.toString(Favorites.ITEM_TYPE_FOLDER) });
+                            + " WHERE container IN (SELECT _id FROM favorites WHERE itemType = ?)"
+                            + " GROUP BY container;",
+                    new String[]{Integer.toString(Favorites.ITEM_TYPE_FOLDER)});
 
             while (c.moveToNext()) {
                 db.execSQL("UPDATE favorites SET rank=cellX+(cellY*?) WHERE "
-                        + "container=? AND cellX IS NOT NULL AND cellY IS NOT NULL;",
-                        new Object[] { c.getLong(1) + 1, c.getLong(0) });
+                                + "container=? AND cellX IS NOT NULL AND cellY IS NOT NULL;",
+                        new Object[]{c.getLong(1) + 1, c.getLong(0)});
             }
 
             c.close();
@@ -449,14 +442,10 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
         return true;
     }
 
-    // Generates a new ID to use for an object in your database. This method should
-    // be only
-    // called from the main UI thread. As an exception, we do call it when we call
-    // the
-    // constructor from the worker thread; however, this doesn't extend until after
-    // the
-    // constructor is called, and we only pass a reference to LauncherProvider to
-    // LauncherApp
+    // Generates a new ID to use for an object in your database. This method should be only
+    // called from the main UI thread. As an exception, we do call it when we call the
+    // constructor from the worker thread; however, this doesn't extend until after the
+    // constructor is called, and we only pass a reference to LauncherProvider to LauncherApp
     // after that point
     @Override
     public int generateNewItemId() {
@@ -502,8 +491,7 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
     }
 
     /**
-     * Returns a new ID to use for a workspace screen in your database that is
-     * greater than all
+     * Returns a new ID to use for a workspace screen in your database that is greater than all
      * existing screen IDs
      */
     public int getNewScreenId() {
@@ -536,8 +524,7 @@ public class DatabaseHelper extends NoLocaleSQLiteHelper implements
         } catch (IllegalArgumentException exception) {
             String message = exception.getMessage();
             if (message.contains("re-open") && message.contains("already-closed")) {
-                // Don't crash trying to end a transaction an an already closed DB. See
-                // b/173162852.
+                // Don't crash trying to end a transaction an an already closed DB. See b/173162852.
             } else {
                 throw exception;
             }
